scripts/aws/test/requirements.txt
pytest>=8.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-jose[cryptography]>=3.3.0
